Move AzureSynapsePipelineHook class to synapse.py for ease of find     * Fix all imports for the class
    * Remove the file from provider.yaml

    * Delete the synapse_pipeline.py file
